+++ MovieTicket.Application/Services/UserService.cs
@@ -8,51 +8,52 @@ using MovieTicket.Domain.Interfaces;
 
 #endregion
 
-namespace MovieTicket.Application.Services;
-
-public class UserService : IUserService
+namespace MovieTicket.Application.Services
 {
-    private readonly IMapper _mapper;
-    private readonly IUserRepository _userRepository;
-
-    public UserService(IUserRepository userRepository,
-        IMapper mapper)
+    public class UserService : IUserService
     {
-        _userRepository = userRepository;
-        _mapper = mapper;
-    }
+        private readonly IMapper _mapper;
+        private readonly IUserRepository _userRepository;
 
-    public async Task<IEnumerable<UserDto>> GetUsers()
-    {
-        var users = await _userRepository.GetUsersAsync();
-        var usersDto = _mapper.Map<IEnumerable<UserDto>>(users);
-        return usersDto;
-    }
+        public UserService(IUserRepository userRepository, IMapper mapper)
+        {
+            _userRepository = userRepository;
+            _mapper = mapper;
+        }
 
-    public async Task<UserDto> GetUserById(int id)
-    {
-        var user = await _userRepository.GetUserByIdAsync(id);
-        var userDto = _mapper.Map<UserDto>(user);
-        return userDto;
-    }
+        public async Task<IEnumerable<UserDto>> GetUsers()
+        {
+            var users = await _userRepository.GetUsersAsync();
+            var usersDto = _mapper.Map<IEnumerable<UserDto>>(users);
+            return usersDto;
+        }
 
-    public async Task<UserDto> AddUser(UserDto userDto)
-    {
-        var user = _mapper.Map<User>(userDto);
-        var newUser = await _userRepository.InsertUserAsync(user);
-        var newUserDto = _mapper.Map<UserDto>(newUser);
-        return newUserDto;
-    }
+        public async Task<UserDto> GetUserById(int id)
+        {
+            var user = await _userRepository.GetUserByIdAsync(id);
+            var userDto = _mapper.Map<UserDto>(user);
+            return userDto;
+        }
 
-    public async Task UpdateUser(UserDto userDto)
-    {
-        var user = _mapper.Map<User>(userDto);
-        await _userRepository.UpdateUserAsync(user);
-    }
+        public async Task<UserDto> AddUser(UserDto userDto)
+        {
+            var user = _mapper.Map<User>(userDto);
+            var newUser = await _userRepository.InsertUserAsync(user);
+            var newUserDto = _mapper.Map<UserDto>(newUser);
+            return newUserDto;
+        }
 
-    public async Task RemoveUser(int id)
-    {
-        var user = _userRepository.GetUserByIdAsync(id).Result;
-        await _userRepository.DeleteUserAsync(user);
+        public async Task UpdateUser(UserDto userDto, int id)
+        {
+            var user = _mapper.Map<User>(userDto);
+            await _userRepository.UpdateUserAsync(user, id);
+        }
+
+        public async Task RemoveUser(int id)
+        {
+            var user = await _userRepository.GetUserByIdAsync(id);
+
+            if (user != null) await _userRepository.DeleteUserAsync(user);
+        }
     }
 }
